requirements.txt
fastapi
uvicorn
pydantic
requests
pandas
httpx
pytesseract
pillow
speechrecognition
markdown
gitpython
beautifulsoup4
numpy
python-dateutil
scipy
pybase64
python-dotenv
markdown
duckdb
openai
gitpython